+++ molgenis-api-permissions/src/main/java/org/molgenis/api/permissions/inheritance/PermissionInheritanceResolver.java
@@ -1,7 +1,7 @@
 package org.molgenis.api.permissions.inheritance;
 
 import static java.util.Objects.requireNonNull;
-import static org.molgenis.api.permissions.PermissionsApiService.ENTITY_PREFIX;
+import static org.molgenis.api.permissions.PermissionsApiServiceImpl.ENTITY_PREFIX;
 import static org.molgenis.api.permissions.UserRoleTools.getRole;
 
 import java.util.ArrayList;
