+++ molgenis-api-permissions/src/test/java/org/molgenis/api/permissions/inheritance/PermissionInheritanceResolverTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.molgenis.api.permissions.inheritance.InheritanceTestUtils.getInheritedPermissionsResult;
-import static org.testng.Assert.*;
+import static org.testng.Assert.assertEquals;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -16,23 +16,29 @@ import org.molgenis.api.permissions.UserRoleTools;
 import org.molgenis.api.permissions.exceptions.InvalidTypeIdException;
 import org.molgenis.api.permissions.inheritance.model.InheritedPermissionsResult;
 import org.molgenis.api.permissions.model.response.InheritedPermission;
-import org.molgenis.data.AbstractMolgenisSpringTest;
 import org.molgenis.data.DataService;
 import org.molgenis.data.Entity;
 import org.molgenis.data.Repository;
 import org.molgenis.data.meta.model.EntityType;
+import org.molgenis.test.AbstractMockitoTest;
 import org.springframework.security.acls.domain.GrantedAuthoritySid;
 import org.springframework.security.acls.domain.PrincipalSid;
 import org.springframework.security.acls.model.Acl;
 import org.springframework.security.acls.model.ObjectIdentity;
 import org.springframework.security.acls.model.Sid;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 // Check InheritanceTestUtils for a description of the test setup
-public class PermissionInheritanceResolverTest extends AbstractMolgenisSpringTest {
+public class PermissionInheritanceResolverTest extends AbstractMockitoTest {
+  @Mock private DataService dataService;
+  @Mock private UserRoleTools userRoleTools;
+  private PermissionInheritanceResolver resolver;
 
-  @Mock DataService dataService;
-  @Mock UserRoleTools userRoleTools;
+  @BeforeMethod
+  private void setUpBeforeMethod() {
+    resolver = new PermissionInheritanceResolver(dataService, userRoleTools);
+  }
 
   @Test
   public void testGetInheritedPermissions() {
@@ -55,8 +61,7 @@ public class PermissionInheritanceResolverTest extends AbstractMolgenisSpringTes
     InheritedPermissionsResult expected =
         getInheritedPermissionsResult(
             entityAcl, packageAcl, parentPackageAcl, role1Sid, role2Sid, role3Sid);
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
+
     assertEquals(resolver.getInheritedPermissions(entityAcl, user), expected);
   }
 
@@ -80,7 +85,6 @@ public class PermissionInheritanceResolverTest extends AbstractMolgenisSpringTes
     InheritedPermissionsResult input =
         getInheritedPermissionsResult(
             entityAcl, packageAcl, parentPackageAcl, role1Sid, role2Sid, role3Sid);
-    DataService dataService = mock(DataService.class);
     @SuppressWarnings("unchecked")
     Repository<Entity> packageRepo = mock(Repository.class);
 
@@ -95,8 +99,6 @@ public class PermissionInheritanceResolverTest extends AbstractMolgenisSpringTes
     doReturn(parent).when(packageRepo).findOneById("parent");
     doReturn(packageRepo).when(dataService).getRepository("sys_md_Package");
 
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     List<InheritedPermission> actual = resolver.convertToInheritedPermissions(input);
 
     // expected
@@ -123,43 +125,31 @@ public class PermissionInheritanceResolverTest extends AbstractMolgenisSpringTes
 
   @Test
   public void testGetEntityTypeIdFromClassPlugin() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("plugin"), "sys_Plugin");
   }
 
   @Test
   public void testGetEntityTypeIdFromClassET() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("entityType"), "sys_md_EntityType");
   }
 
   @Test
   public void testGetEntityTypeIdFromClassGroup() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("group"), "sys_sec_Group");
   }
 
   @Test
   public void testGetEntityTypeIdFromClassPack() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("package"), "sys_md_Package");
   }
 
   @Test
   public void testGetEntityTypeIdFromClassEntity() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("entity-test"), "test");
   }
 
   @Test(expectedExceptions = InvalidTypeIdException.class)
   public void testGetEntityTypeIdFromClassError() {
-    PermissionInheritanceResolver resolver =
-        new PermissionInheritanceResolver(dataService, userRoleTools);
     assertEquals(resolver.getEntityTypeIdFromClass("error"), "");
   }
 }
